backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pydantic[email]>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
mysql-connector-python>=8.0.0
pymysql>=1.1.0
